% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif.R
\name{classifTypicalDay}
\alias{classifTypicalDay}
\title{Generate clusturing of typical day}
\usage{
classifTypicalDay(calendar, vertices, nbClustWeek = 3, nbClustWeekend = 1,
  report = TRUE, reportPath = setwd())
}
\arguments{
\item{calendar}{\code{list}, vector of date in eatch period can be obtain which \link{getCalendar}}

\item{vertices}{\code{data.table}, 5 column :
\itemize{
 \item Date : date (%Y-%M-%D)
 \item Period : Hour, (1:24)
 \item BE : Belgium vertices
 \item DE : deutschland vertices
 \item FR : french vertices
}}

\item{nbClustWeek}{\code{numeric} number of cluster for week period}

\item{nbClustWeekend}{\code{numeric} number of cluster for weekend period}

\item{report}{\code{boolean}, generate reports}

\item{reportPath}{\code{character}, patch for reports generation}
}
\description{
Generate clusturing of typical day
}
\examples{

\dontrun{
library(data.table)
vertices <- fread(system.file("dev/verticesAllDay.txt",package = "flowBasedClustering"))
dates <- generateAllDate("2015-11-01", "2017-01-20")
interSeasonBegin <- c("2016-03-01", "2016-10-01")
interSeasonEnd <- c("2016-05-15", "2016-10-31")
dayInWeekend = c(6, 7)
calendar <- getCalendar(dates, interSeasonBegin, interSeasonEnd)
myClassif <- classifTypicalDay(calendar, vertices)
}


}
