% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringTypicalDays.R
\name{clusteringTypicalDays}
\alias{clusteringTypicalDays}
\title{Generate clustering of typical days}
\usage{
clusteringTypicalDays(calendar, vertices, nbClustWeek = 3,
  nbClustWeekend = 1, report = TRUE, reportPath = getwd(),
  hourWeigth = rep(1, 24))
}
\arguments{
\item{calendar}{\code{list}, vector of date for each period. Can be obtain with \link{getCalendar}}

\item{vertices}{\code{data.table}, 5 columns :
\itemize{
 \item Date : date (%Y-%M-%D)
 \item Period : Hour (1:24)
 \item BE : belgium vertices
 \item DE : german vertices
 \item FR : french vertices
}
This parameter can be obtained with the function \code{ptdfToVertices()}.}

\item{nbClustWeek}{\code{numeric}, number of clusters for week (working days) period. Defaut to 3}

\item{nbClustWeekend}{\code{numeric}, number of clusters for weekend period. Defaut to 1}

\item{report}{\code{boolean}, generate report. Defaut to TRUE}

\item{reportPath}{\code{character}, path where the report is written. Defaut to \code{getwd()}}

\item{hourWeigth}{\code{numeric}, vector of 24 weights to ponderate differently each hour of the day}
}
\description{
Generate clustering of typical days
}
\examples{

\dontrun{
library(data.table)
vertices <- fread(system.file("dev/verticesAllDay.txt",package = "flowBasedClustering"))

dates <- getSequence("2015-11-01", "2017-01-20")
interSeasonBegin <- c("2016-03-01", "2016-10-01")
interSeasonEnd <- c("2016-05-15", "2016-10-31")
calendar <- getCalendar(dates, interSeasonBegin, interSeasonEnd)

clusterTD <- clusteringTypicalDays(calendar, vertices)
}


}

