% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTypicalDaysForOneClass.R
\name{clusterTypicalDaysForOneClass}
\alias{clusterTypicalDaysForOneClass}
\title{Generate a set of flow-based typical days for a season}
\usage{
clusterTypicalDaysForOneClass(dates, vertices, nbCluster, hourWeight = rep(1,
  24), className = NULL, reportPath = getwd(), report = TRUE,
  id_start = 1)
}
\arguments{
\item{dates}{\code{vector}, vector of date.}

\item{vertices}{\code{data.table}, 5 columns :
\itemize{
 \item Date : date (%Y-%M-%D)
 \item Period : Hour (1:24)
 \item BE : belgium vertices
 \item DE : german vertices
 \item FR : french vertices
}
This parameter can be obtained with the function \code{ptdfToVertices()}.}

\item{nbCluster}{\code{numeric}, number of clusters.}

\item{hourWeight}{\code{numeric}, vector of 24 weights to ponderate differently each hour of the day}

\item{className}{\code{character}, name of class.}

\item{reportPath}{\code{character}, path where the report is written. Defaut to \code{getwd()}}

\item{report}{\code{boolean}, if TRUE, reports are generated.}

\item{id_start}{\code{numeric}, first id of typical days.}
}
\description{
To do
}
\examples{

\dontrun{
library(data.table) 
vertices <- fread(system.file("dataset/vertices_example.txt",package = "flowBasedClustering"))
dates <- seq(as.Date("2016-08-09"), as.Date("2018-09-01"), by = "day")
hourWeight = rep(1, 24)

clusterTypicalDaysForOneClass(vertices = vertices,
                              dates = dates, nbCluster = 2,
                              className = "myName", id_start = 5)
}


}
