% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPtdfForCompute.R
\name{createPtdfForCompute}
\alias{createPtdfForCompute}
\title{Transform a PTDF file}
\usage{
createPtdfForCompute(cluster, PTDF = system.file("dataset/ptdf_example.csv",
  package = "flowBasedClustering"), pathOutput = getwd())
}
\arguments{
\item{cluster}{\code{data.table} output of \link{clusteringTypicalDays}}

\item{PTDF}{\code{character}, path of a PTDF file. 
The PTDF file must contains at least these seven columns : Date, Period, BE, DE, FR, NL and RAM - in this order.
It must be a .csv file with ";" as column separator and "." as decimal separator.
The names of the seven columns (Date, Period, ...) must be included in the header (first line)
of the .csv file.}

\item{pathOutput}{\code{character} path of folder where file will be write, default getwd()}
}
\description{
Link function to transform a PTDF file for computeFB function (package antaresFlowbased)
}
\examples{

\dontrun{
#load clustering results (or build them with clusteringTypicalDays function())
clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))

PTDF = system.file("dataset/ptdf_example.csv",
      package = "flowBasedClustering")
      
createPtdfForCompute(clusterTD, PTDF)
#Write on D:/ :
createPtdfForCompute(clusterTD, PTDF, pathOutput = "D:/")

}

}
