% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendar.R
\name{cutYear}
\alias{cutYear}
\title{Define a calender which season}
\usage{
cutYear(dates, interSeasonBegin, interSeasonEnd, dayExclude = NULL,
  holiday = c("NewYearsDay", "GoodFriday", "EasterSunday", "EasterMonday",
  "LaborDay", "FRAscension", "Pentecost", "PentecostMonday",
  "FRAssumptionVirginMary", "FRAllSaints", "FRArmisticeDay", "ChristmasEve",
  "ChristmasDay", "BoxingDay", "DENewYearsEve"), dayInWeekend = c(6, 7))
}
\arguments{
\item{dates}{\code{character or date}, vector of dates YYYY-MM-DD, can be generate which \link{generateAllDate}}

\item{interSeasonBegin}{\code{character or date}, date, YYYY-MM-DD, begin of interSaison}

\item{interSeasonEnd}{\code{character or date}, date, YYYY-MM-DD, end of interSaison}

\item{dayExclude}{\code{character or date}, date, YYYY-MM-DD, day to exclude}

\item{holiday}{\code{character}, holiday day default :
NewYearsDay,GoodFriday,EasterSunday,
EasterMonday,LaborDay,FRAscension,Pentecost,PentecostMonday,
FRAssumptionVirginMary,FRAllSaints,FRArmisticeDay,
ChristmasEve,ChristmasDay,BoxingDay,DENewYearsEve}

\item{dayInWeekend}{\code{numeric}, indice of weekend day}
}
\description{
Define a calender which season
}
\examples{

\dontrun{
dates <- generateAllDate("2017-01-01", "2017-12-31")
interSeasonBegin <- c("2017-03-01", "2017-09-01")
interSeasonEnd <- c("2017-05-01", "2017-11-01")
dayInWeekend = c(6, 7)
cutYear(dates, interSeasonBegin, interSeasonEnd)
}
}
