% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromptdftopolyedron.R
\name{fromPtdfToVertices}
\alias{fromPtdfToVertices}
\title{Transforme a PTDF file to vertives file}
\usage{
fromPtdfToVertices(PTDF = system.file("dav/data/faceAllYear.csv", package =
  "flowBasedClustering"), nbCors = 1)
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file}

\item{nbCors}{\code{numeric}, use parallelisation or not.Indicate number of cores use. If 1 (default) a no parallel.}
}
\description{
Transforme a PTDF file to vertives file
}
\examples{

\dontrun{
fromPtdfToVertices(nbCors = 4)
}

}
