% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendar.R
\name{generateAllDate}
\alias{generateAllDate}
\title{Create a vector of date from a date begin and a date end}
\usage{
generateAllDate(dateBegin, dateEnd)
}
\arguments{
\item{dateBegin}{\code{character or date}, date, YYYY-MM-DD}

\item{dateEnd}{\code{character or date}, date, YYYY-MM-DD}
}
\description{
Create a vector of date from a date begin and a date end
}
\examples{

\dontrun{
generateAllDate("2017-01-01", "2017-12-31")
}

}
