% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateRaportClustering}
\alias{generateRaportClustering}
\title{Generate html report}
\usage{
generateRaportClustering(dayType, output_file = NULL, data = NULL)
}
\arguments{
\item{dayType}{\code{numeric} Typical day}

\item{output_file}{\code{character} path of output}

\item{data}{\code{data.table} data from \link{classifTipycalDay}}
}
\description{
Generate html report
}
\examples{

\dontrun{
generateRaportClustering(dayType = 7)
}
}
