% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendar.R
\name{getCalendar}
\alias{getCalendar}
\title{Define a calendar with season}
\usage{
getCalendar(dates, interSeasonBegin, interSeasonEnd, dayExclude = NULL,
  holiday = c("NewYearsDay", "GoodFriday", "EasterSunday", "EasterMonday",
  "LaborDay", "FRAscension", "Pentecost", "PentecostMonday",
  "FRAssumptionVirginMary", "FRAllSaints", "FRArmisticeDay", "ChristmasEve",
  "ChristmasDay", "BoxingDay", "DENewYearsEve"), dayInWeekend = c(6, 7))
}
\arguments{
\item{dates}{\code{character or date}, vector of dates YYYY-MM-DD, can be generate with \link{getSequence}}

\item{interSeasonBegin}{\code{character or date}, date, YYYY-MM-DD, begin of interseason}

\item{interSeasonEnd}{\code{character or date}, date, YYYY-MM-DD, end of interseason}

\item{dayExclude}{\code{character or date}, date, YYYY-MM-DD, day to exclude}

\item{holiday}{\code{character}, see \link[timeDate]{holiday} for holidays names, holidays days default :
NewYearsDay,GoodFriday,EasterSunday,
EasterMonday,LaborDay,FRAscension,Pentecost,PentecostMonday,
FRAssumptionVirginMary,FRAllSaints,FRArmisticeDay,
ChristmasEve,ChristmasDay,BoxingDay,DENewYearsEve}

\item{dayInWeekend}{\code{numeric}, indice of weekend day}
}
\description{
Define a calendar with season
}
\examples{

\dontrun{
dates <- getSequence("2017-01-01", "2017-12-31")
interSeasonBegin <- c("2017-03-01", "2017-09-01")
interSeasonEnd <- c("2017-05-01", "2017-11-01")
dayExclude <- dates
getCalendar(dates, interSeasonBegin, interSeasonEnd, dayExclude = dayExclude)
}

}

