% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{getProbability}
\alias{getProbability}
\title{Get probability and quantiles from climat file and classif}
\usage{
getProbability(climat, classif)
}
\arguments{
\item{climat}{\code{data.table} climat file, 2 first columns are Date and Period, others are params files}

\item{classif}{\code{data.table} output of \link{classifTypicalDay}}
}
\description{
Get probability and quantiles from climat file and classif
}
\examples{

\dontrun{
climat <- fread(system.file("dev/Climat.txt",package = "flowBasedClustering"))
classif <- readRDS(system.file("dev/ClassifOut.RDS",package = "flowBasedClustering"))

MatProb <- getProbability(climat, classif)
}
}
