% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendar.R
\name{getSequence}
\alias{getSequence}
\title{Create a vector of date from a being date begin to a end date}
\usage{
getSequence(dateBegin, dateEnd)
}
\arguments{
\item{dateBegin}{\code{character or date}, date, YYYY-MM-DD}

\item{dateEnd}{\code{character or date}, date, YYYY-MM-DD}
}
\description{
Create a vector of date from a being date begin to a end date
}
\examples{

\dontrun{
getSequence("2017-01-01", "2017-12-31")
}

}

