% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plotFlowbased}
\alias{plotFlowbased}
\title{Plot flow-based domain(s)}
\usage{
plotFlowbased(PTDF, country1, country2, domainsNames = NULL, main = "",
  xlim = c(-10000, 10000), ylim = c(-10000, 10000), width = "420px",
  height = "410px")
}
\arguments{
\item{PTDF}{\code{data.frame or list} (list of) data.frame(s) with the PTDF of one time step. 
The data.frame must have the same columns as the input file of \link{ptdfToVertices}}

\item{country1}{\code{character}, name of the country whose net position is in the x axis (BE, FR, DE or NL)}

\item{country2}{\code{character}, name of the country whose net position is in the y axis (BE, FR, DE or NL)}

\item{domainsNames}{\code{character} names of the domain(s), used as legend of the graph}

\item{main}{\code{character} title of the graph}

\item{xlim}{\code{numeric} x axis limits, vector with two values (min and max)}

\item{ylim}{\code{numeric} y axis limits, vector with two values (min and max)}

\item{width}{\code{character}, for rAmCharts only. Default to "420px" (set to "100/100" for dynamic resize)}

\item{height}{\code{character}, for rAmCharts only. Default to "410px" (set to "100/100" for dynamic resize)}
}
\description{
Plot flow-based domain(s)
}
\examples{

\dontrun{
PTDF <- fread(system.file("dataset/ptdf_example.csv",package = "flowBasedClustering"))
#Plot unique PTDF
PTDF1 <- PTDF[ Date == "06/08/2016" & Period == 1]
plotFlowbased(PTDF1, country1 = "DE", country2 = "BE", domainsNames = "06/08/2016")

PTDF2 <- list(PTDF[ Date == "06/08/2016" & Period == 1],
PTDF[ Date == "06/08/2016" & Period == 2],
PTDF[ Date == "06/08/2016" & Period == 3])
plotFlowbased(PTDF2, country1 = "DE", country2 = "BE", domainsNames = c("01/11/2015 - H1",
"01/11/2015 - H2", "01/11/2015 - H3"), main = "Myplot")

}

}
