% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptdfToVertices.R
\name{ptdfToVertices}
\alias{ptdfToVertices}
\title{Convert a PTDF file into a table of vertices}
\usage{
ptdfToVertices(PTDF = system.file("dataset/ptdf_example.csv", package =
  "flowBasedClustering"), nbCore = 1)
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file. 
The PTDF file must contains at least these seven columns Date, Period, BE, DE, FR, NL, RAM in this order.}

\item{nbCore}{\code{numeric}, number of cores for parallel computation. Default to one}
}
\description{
PTDF are the equations of the hyperplanes defining the limits of flow-based domains, 
while the vertices are the coordinnates of the extreme points of the domains.
}
\examples{

\dontrun{
ptdf_data <- data.table::fread(system.file("dataset/ptdf_example.csv",package = "flowBasedClustering"),
data.table = F)

head(ptdf_data[, 1:7])

vertices <- data.table::fread(system.file("dataset/vertices_example.txt",package = "flowBasedClustering"))

head(vertices)

}

}
