% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Generate a plot for a typical day cluster}
\usage{
clusterPlot(data, country1, country2, hour, dayType)
}
\arguments{
\item{data}{\code{data.table}, output of \link{classifTipycalDay}}

\item{country1}{\code{character}, name of first country}

\item{country2}{\code{character}, name of second country}

\item{hour}{\code{numeric}, hour}

\item{dayType}{\code{numeric}, dayType}

\item{holiday}{\code{character}, holiday day default :}
}
\description{
Generate a plot for a typical day cluster
}
