% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Plot a flow-based domain of a typical day}
\usage{
clusterPlot(data, country1, country2, hour, dayType, typicalDayOnly = FALSE,
  ggplot = FALSE)
}
\arguments{
\item{data}{\code{data.table} results from the clustering, output data from \link{clusteringTypicalDays}}

\item{country1}{\code{character}, name of the country whose net position is in the x axis}

\item{country2}{\code{character}, name of the country whose net position is in the y axis}

\item{hour}{\code{numeric}, hour of the plotted domain}

\item{dayType}{\code{numeric}, typical flow-based day identifier}

\item{typicalDayOnly}{\code{logical} if TRUE, plot only the domain of the typical day and not the other domains of the cluster}

\item{ggplot}{\code{logical} should ggplot package be used (static graph) instead of ramCharts (dynamic graph)}
}
\description{
For a given hour, plot the flow-based domain of one typical day along with the other
flow-based domains it represents (i.e. other domains of the same cluster).
}
\examples{

\dontrun{

# classification result
clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))

clusterPlot(clusterTD, "FR", "DE", 8, 9, FALSE, FALSE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, FALSE, TRUE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, TRUE, TRUE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, TRUE, FALSE)

}
}
