% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{clusterPlot}
\alias{clusterPlot}
\title{Generate a plot for a typical day cluster}
\usage{
clusterPlot(data, country1, country2, hour, dayType, typicalDayOnly = FALSE,
  ggplot = FALSE)
}
\arguments{
\item{data}{\code{data.table}, output of \link{clusteringTypicalDays}}

\item{country1}{\code{character}, name of first country}

\item{country2}{\code{character}, name of second country}

\item{hour}{\code{numeric}, hour}

\item{dayType}{\code{numeric}, dayType}

\item{typicalDayOnly}{: plot only typical day ?}

\item{ggplot}{: ggplot or amCharts ?}
}
\description{
Generate a plot for a typical day cluster
}
\examples{

\dontrun{

# classification result
clusterTD <- readRDS(system.file("dev/ClassifOut.RDS",package = "flowBasedClustering"))

clusterPlot(clusterTD, "FR", "DE", 8, 9, FALSE, FALSE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, FALSE, TRUE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, TRUE, TRUE)
clusterPlot(clusterTD, "FR", "DE", 8, 9, TRUE, FALSE)

}
}
