% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTypicalDaysForOneClass.R
\name{clusterTypicalDaysForOneClass}
\alias{clusterTypicalDaysForOneClass}
\title{Generate a set of flow-based typical days on one time period}
\usage{
clusterTypicalDaysForOneClass(dates, vertices, nbCluster, hourWeight = rep(1,
  24), className = NULL, reportPath = getwd(), report = TRUE,
  id_start = 1, maxDomainSize = 10000)
}
\arguments{
\item{dates}{\code{character}, vector of date (format YYYY-MM-DD).}

\item{vertices}{\code{data.table}, 5 columns :
\itemize{
 \item Date : date (%Y-%M-%D)
 \item Period : Hour (1:24), period i corresponds to the hourly time-step [i-1 ; i]
 \item BE : belgium vertices
 \item DE : german vertices
 \item FR : french vertices
}
This parameter can be obtained with the function \link{ptdfToVertices}.}

\item{nbCluster}{\code{numeric}, number of clusters}

\item{hourWeight}{\code{numeric}, vector of 24 weights, one for each hour of the day. The clustering algorithm
will be more accurate for the flow-based domains of the hours with a relatively higher weight.}

\item{className}{\code{character}, name of the class characterizing the studied time period}

\item{reportPath}{\code{character}, path where the report is written. Defaut to \code{getwd()}}

\item{report}{\code{boolean}, if TRUE, reports are generated.}

\item{id_start}{\code{numeric}, first identifier of the returned typical days. Default value is 1}

\item{maxDomainSize}{\code{numeric} limit of domain size in each axis. The function will return an error if one domain
or more exceed these limits.}
}
\description{
Run a clustering algorithm for a given historical period. It creates clusters by
gathering the most similar days and chooses among them the best
representative: it will be a so-called typical day. The metric used to determine the similarity of two days is
a weighted sum of 24 hourly distances, meaning the distances between the domains of the two
days at the same hour.
}
\examples{

\dontrun{
library(data.table)
vertices <- fread(system.file("dataset/vertices_example.txt",package = "flowBasedClustering"))
dates <- seq(as.Date("2016-08-09"), as.Date("2018-09-01"), by = "day")
hourWeight = rep(1, 24)

out1 <- clusterTypicalDaysForOneClass(vertices = vertices,
                                      dates = dates, nbCluster = 2,
                                      className = "myName", id_start = 5, report = FALSE)

out2 <- clusterTypicalDaysForOneClass(vertices = vertices,
                                      dates = dates, nbCluster = 4,
                                      className = "myName2", id_start = 7, report = FALSE)

rbindlist(list(out1, out2))
}


}
