% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateClusteringReport}
\alias{generateClusteringReport}
\title{Generate html report}
\usage{
generateClusteringReport(dayType, outputFile = NULL, data = NULL)
}
\arguments{
\item{dayType}{\code{numeric} Typical day}

\item{outputFile}{\code{character} path of output}

\item{data}{\code{data.table} data from \link{clusteringTypicalDays}}
}
\description{
Generate html report
}
\examples{

\dontrun{
generateClusteringReport(dayType = 7)
}
}

