% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateClusteringReport}
\alias{generateClusteringReport}
\title{Generate html report}
\usage{
generateClusteringReport(dayType, outputFile = NULL, data)
}
\arguments{
\item{dayType}{\code{numeric} Typical day identifier}

\item{outputFile}{\code{character} a folder where the html report is saved}

\item{data}{\code{data.table} results from the clustering, output data from \link{clusteringTypicalDays}}
}
\description{
Write and save a report with the results of the clustering of the flow-based domains. For one typical day,
the report contains the representation of all the 24 hourly flow-based domains and their comparison
with the domains of the other historical days it represents.
}
\examples{

\donttest{

clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))
generateClusteringReport(dayType = 7, data = clusterTD)

}
}
