% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R, R/graphs.R
\name{generateClusteringReport}
\alias{generateClusteringReport}
\alias{generateClusteringReport}
\title{Generate html report}
\usage{
generateClusteringReport(dayType, outputPath = NULL, data = NULL)

generateClusteringReport(dayType, outputPath = NULL, data = NULL)
}
\arguments{
\item{dayType}{\code{numeric} Typical day}

\item{outputPath}{\code{character} path of output}

\item{data}{\code{data.table} data from \link{classifTypicalDay}}

\item{output_file}{\code{character} path of output}

\item{dayType}{\code{numeric} Typical day}

\item{data}{\code{data.table} output of \link{clusteringTypicalDays}}
}
\description{
Generate html report

Generate html report
}
\examples{

\dontrun{
generateClusteringReport(dayType = 7)
}

\dontrun{
# classification result
clusterTD <- readRDS(system.file("dev/ClassifOut.RDS",package = "flowBasedClustering"))

generateClusteringReport(dayType = 7, data = clusterTD)

}

}
