% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{getProbability}
\alias{getProbability}
\title{Get probabilities and quantiles from climate file for a classification}
\usage{
getProbability(climate, cluster, levelsProba = c(0.333, 0.666),
  extrapolationNA = TRUE)
}
\arguments{
\item{climate}{\code{data.table} climate file, 2 first columns are Date and Period, others are params files}

\item{cluster}{\code{data.table} output of \link{clusteringTypicalDays}}

\item{levelsProba}{\code{numeric or list}, can be \code{numeric} or \code{list}, if \code{numeric}, all quantiles
are the same for all climate variables, if \code{list}, must be names and give quantiles for each climate variable.}

\item{extrapolationNA}{\code{booelan} extrapolate NA value wich neighbours.}
}
\description{
Get probabilities and quantiles from climate file for a classification
}
\examples{

\dontrun{
# climate data
climate <- fread(system.file("dev/climat.txt",package = "flowBasedClustering"))

# classification result
clusterTD <- readRDS(system.file("dev/ClassifOut.RDS",package = "flowBasedClustering"))

levelsProba <- list(`Conso (J-1)` = c(0.2, 0.4, 0.8), DE_wind = c(0.3, 0.7),DE_solar = c(0.3,0.4,0.8) )
MatProb <- getProbability(climate, clusterTD, levelsProba = levelsProba)

levelsProba <- list(`Conso (J-1)` = c(0.2,0.25,0.5,0.75,0.90, 0.95), DE_wind = c(0.25,0.5,0.75, 0.9, 0.95),DE_solar = c(0.25,0.5,0.75) )
MatProb <- getProbability(climate, clusterTD, levelsProba = levelsProba, extrapolationNA = TRUE)
}
}
