% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{getProbability}
\alias{getProbability}
\title{Get probabilities and quantiles from climate file for a classification}
\usage{
getProbability(climate, cluster, levelsProba = c(1/3, 2/3),
  extrapolationNA = TRUE)
}
\arguments{
\item{climate}{\code{data.table} climate file, 2 first columns are Date and Period, others are params files}

\item{cluster}{\code{data.table} output of \link{clusteringTypicalDays}}

\item{levelsProba}{\code{numeric or list}, can be \code{numeric} or \code{list}, if \code{numeric}, all quantiles
are the same for all climate variables, if \code{list}, must be names and give quantiles for each climate variable.}

\item{extrapolationNA}{\code{booelan} extrapolate NA value wich neighbours.}
}
\description{
Get probabilities and quantiles from climate file for a classification
}
\examples{

\dontrun{
# load climate daily time serires
climate <- fread(system.file("dataset/climate_example.txt",package = "flowBasedClustering"))

# load clustering results (or build them with clusteringTypicalDays function())
clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))

# get Probability matrices
MatProb <- getProbability(climate, clusterTD)

# run with specified probability levels, and witout auto-filling the NA cases
levelsProba <- list(FR_load = c(0.2, 0.4, 0.8), DE_wind = c(0.3, 0.7),DE_solar = c(0.3,0.4,0.8) )
MatProb <- getProbability(climate, clusterTD, levelsProba = levelsProba, extrapolationNA = FALSE)

}
}
