% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalendar.R
\name{getSequence}
\alias{getSequence}
\title{Return a vector of consecutive dates}
\usage{
getSequence(dateBegin, dateEnd)
}
\arguments{
\item{dateBegin}{\code{character or date}, date, YYYY-MM-DD}

\item{dateEnd}{\code{character or date}, date, YYYY-MM-DD}
}
\value{
a vector of dates
}
\description{
Return a vector of consecutive dates
}
\examples{

\dontrun{
getSequence("2017-01-01", "2017-12-31")
}

}
