% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVertices.R
\name{getVertices}
\alias{getVertices}
\title{Get vertices from faces}
\usage{
getVertices(face, b)
}
\arguments{
\item{face}{\code{data.table}, face for 3 countries (BE, DE and FR)}

\item{b}{\code{numeric}, extreme point b}
}
\description{
Get vertices from faces
}

