% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plotFlowbased}
\alias{plotFlowbased}
\title{Plot flow-based domain(s)}
\usage{
plotFlowbased(PTDF, country1, country2, domainsNames = NULL, main = "")
}
\arguments{
\item{PTDF}{\code{data.frame or list} PTDF file}

\item{country1}{\code{character} country 1 possible values are : BE, DE, FR, NL}

\item{country2}{\code{character} country 2 possible values are : BE, DE, FR, NL}

\item{domainsNames}{\code{character} names for domain(s), use in legend}

\item{main}{\code{character} title}
}
\description{
Plot flow-based domain(s)
}
\examples{

\dontrun{
PTDF <- fread(system.file("dev/data/faceAllYear.csv",package = "flowBasedClustering"))
#Plot unique PTDF
PTDF1 <- PTDF[ Date == "01/11/2015" & Period == 1]
plotFlowbased(PTDF1, country1 = "DE", country2 = "BE", domainsNames = "01/11/2015")

PTDF2 <- list(PTDF[ Date == "01/11/2015" & Period == 1],
PTDF[ Date == "01/11/2015" & Period == 2],
PTDF[ Date == "01/11/2015" & Period == 3])
plotFlowbased(PTDF2, country1 = "DE", country2 = "BE", domainsNames = c("01/11/2015 - H1",
"01/11/2015 - H2", "01/11/2015 - H3"), main = "Myplot")

}

}

