% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{plotMonotone}
\alias{plotMonotone}
\title{Plot monotone}
\usage{
plotMonotone(climat, classif, hour, dayType, variable)
}
\arguments{
\item{climat}{\code{data.table} climat file, 2 first columns are Date and Period, others are params files}

\item{classif}{\code{data.table} output of \link{classifTipycalDay}}

\item{hour}{\code{numeric} hour}

\item{dayType}{\code{numeric} typical day}

\item{variable}{\code{character} name of variable to plot}
}
\description{
Plot monotone
}
\examples{

\dontrun{
climat <- fread("D:/Users/titorobe/Desktop/Antares/flowBasedClustering/inst/dev/Climat.txt")
classif <- readRDS("D:/Users/titorobe/Desktop/Antares/flowBasedClustering/inst/dev/ClassifOut.RDS")

plotMonotone(climat = climat, classif = classif, hour = 1, dayType = 1, variable = "DE_wind")
}
}
