% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{plotMonotone}
\alias{plotMonotone}
\title{Plot monotone}
\usage{
plotMonotone(climat, classif, dayType, variable)
}
\arguments{
\item{climat}{\code{data.table} climat file, 2 first columns are Date and Period, others are params files}

\item{classif}{\code{data.table} output of \link{clusteringTypicalDays}}

\item{dayType}{\code{numeric} typical day}

\item{variable}{\code{character} name of variable to plot}
}
\description{
Plot monotone
}
\examples{

\dontrun{
# climat data
climat <- fread(system.file("dev/Climat.txt",package = "flowBasedClustering"))

# classification result
clusterTD <- readRDS(system.file("dev/ClassifOut.RDS",package = "flowBasedClustering"))

plotMonotone(climat = climat, classif = clusterTD, dayType = 1, variable = "DE_wind")
}
}
