% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbability.R
\name{plotMonotone}
\alias{plotMonotone}
\title{Plot monotone}
\usage{
plotMonotone(climate, cluster, dayType, variable)
}
\arguments{
\item{climate}{\code{data.table} climate file, 2 first columns are Date and Period, others are params files}

\item{cluster}{\code{data.table} output of \link{clusteringTypicalDays}}

\item{dayType}{\code{numeric} typical day}

\item{variable}{\code{character} name of variable to plot}
}
\description{
Plot monotone
}
\examples{

\dontrun{
# climate data
climate <- fread(system.file("dataset/climate_example.txt",package = "flowBasedClustering"))

# classification result
clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))

plotMonotone(climate = climate, cluster = clusterTD, dayType = 1, variable = "DE_wind")
}
}
