% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptdfToVertices.R
\name{ptdfToVertices}
\alias{ptdfToVertices}
\title{Convert a PTDF file into a table of vertices}
\usage{
ptdfToVertices(PTDF, nbCore = 1, maxDomainSize = 10000)
}
\arguments{
\item{PTDF}{\code{character}, path of a PTDF file. 
The PTDF file must contains at least these seven columns : Date, Period, BE, DE, FR, NL and RAM - in this order.
It must be a .csv file with ";" as column separator and "." as decimal separator.
The names of the seven columns (Date, Period, ...) must be included in the header (first line)
of the .csv file.}

\item{nbCore}{\code{numeric}, number of cores for parallel computation. Default to one.}

\item{maxDomainSize}{\code{numeric} limit of domain size in each axis. The function will return a warning if one domain
or more exceed these limits.}
}
\value{
A data table with the vertices of the flow-based domains.
}
\description{
PTDF are the equations of the hyperplanes defining the limits of flow-based domains, 
while the vertices are the coordinnates of the extreme points of the domains.
}
\examples{

\dontrun{

ptdf_data <- data.table::fread(system.file("dataset/ptdf_example.csv", 
   package = "flowBasedClustering"), data.table = F)

head(ptdf_data)

vertices <- ptdfToVertices(system.file("dataset/ptdf_example.csv",
    package = "flowBasedClustering"))

head(vertices)

}

}
