% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptdfToVertices.R
\name{ptdfToVertices}
\alias{ptdfToVertices}
\title{Transform a PTDF file to vertices file}
\usage{
ptdfToVertices(PTDF = system.file("dev/data/faceAllYear.csv", package =
  "flowBasedClustering"), nbCore = 1)
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file. This file must have 7 obligatory columns : Id_day, Period, BE, DE, FR, NL, RAM}

\item{nbCore}{\code{numeric}, use parallel computing or not. Indicate number of cores use. If 1 (default), no parallel.}
}
\description{
Transform a PTDF file to vertices file
}
\examples{

\dontrun{
ptdfToVertices(nbCore = 4)
}

}

